Implement locale and timezone setting
Add SetLocale and SetTimezone methods to givc-admin for changing the system settings.
When set, the value are stored to a configuration file, and reported back to agents
when they register.

Adjust givc-agent to handle the timezone and locale settings repotrted by the admin; use
systemd's localectl, timedatectl and systemctl set-environment to apply the settings. Add
policykit policy to allow locale and timezone setting to the givc user.

Signed-off-by: Santtu Lakkala <santtu.lakkala@unikie.com>