fix(hwid): fix hwid reporting
As the interface is brought up, it will report a random MAC address.
This can cause issues with external queries, depending on the timing.
Checking device "up" still leaves a chance of a wrong MAC.

Checking /sys/class/net/<iface>/carrier also appears to work, but for
simplicity we are using go's net lib to query if the device is
running.

Additional fixes:
- add ethernet interface as automatic alternative (untested atm)
- bump crane to remove warnings

Signed-off-by: Manuel Bluhm <manuel@ssrc.tii.ae>